Minor changes.
git-svn-id: svn://svn.zabbix.com/trunk@2983 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ src/zabbix_agent/logfiles.h
@@ -20,6 +20,6 @@
 #ifndef ZABBIX_LOGFILES_H
 #define ZABBIX_LOGFILES_H
 
-int   process_log(char *filename,int *lastlogsize, char *value);
+int   process_log(char *filename,long *lastlogsize, char *value);
 
 #endif
